fix: Remove broken page reference in index (FIR-36428) (#989)
Our docs currently have a random entry called `Reference sql` in the index. Clicking on it shows the next page about explicit transactions. This PR removes that. I did some digging whether we used to have a page with content for that. It looks like it was a mostly empty page that was removed deliberately in https://github.com/firebolt-analytics/firebolt-docs-staging/pull/842/files.