chore: COPY FROM index access is not allowed for `TYPE=PARQUET` (FIR-44752) (#993)
We decided in https://firebolt-analytics.slack.com/archives/C07NP6YFDQ8/p1757670446926429 that we will not continue to support index access for parquet files. Add this in the docs.

Also removed the random fact that the index supports up to `2^64-1` columns. Our system does not support reading that many columns in a parquet file, even if the index range is that high.